+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Added `FormEmailInput` component.
 * Added `FormTextInput` component.
 * Added `FormNumberInput` component.
+* Added `FormLabel` component.
 
 ## 5.1.0
 Added `NotNullOrFallback` component.